Вынос типов в отдельный файл - 2
